split server.py into views.py, forms.py, and game.py
