+++ app/application/__init__.py
@@ -1,12 +1,15 @@
 import os
 from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
 
 project_dir = os.path.dirname(os.path.abspath(__file__))
 database_file = "sqlite:///{}".format(os.path.join(project_dir, "gamedatabase.db"))
 
-
 app = Flask(__name__)
 app.config["SQLALCHEMY_DATABASE_URI"] = database_file
 app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
+db = SQLAlchemy(app)
 
-import application.server
+import application.views
+import application.forms
+import application.game
